pyproject.toml
[build-system]


requires = ["poetry-core>=1.0.0"] # Spécifie que Poetry est utilisé pour le build
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "projet-ventes-mle" # Nom du package pour PyPI, peut être différent du nom du répertoire
version = "0.1.0"
description = "Projet d'analyse et prédiction des tendances de vente"
authors = ["aliadilanielfayaz <aliadilanie21@gmail.com>","aliadilanielfayaz <aliadilanielfayaz@yahoo.fr>"]
readme = "README.rst" # Ou README.md si vous prévoyez de changer le format

# Indique à Poetry où trouver le package principal.
# Votre structure est /projet_ventes_mle/projet_ventes_mle/ donc c'est correct.
packages = [{include = "projet_ventes_mle"}]

[tool.poetry.dependencies]
python = "^3.8" # Ou la version de Python que vous ciblez (ex: ^3.9, ~3.10)
fastapi = "^0.109.0" # Poetry installera la dernière version compatible avec cette contrainte
uvicorn = {extras = ["standard"], version = "^0.27.0"} # Idem

[tool.poetry.dev-dependencies]
pytest = "^5.2"
psycopg2-binary = "^2.9.9" # Pour la connexion à PostgreSQL
python-dotenv = "^1.0.0" # Pour charger les variables d'environnement depuis un fichier .env
